Make feature_names include bias term again
The reason it is better to include it is that internally, we need
to have the bias name in feature_names, and if we always add it
to externally passed feature_names, it can be a major bottleneck.
This change allows pre-computing feature_names and passing it as
a numpy array.