Improve is_probabilistic_classifier (fix lighting tests)
Try to look inside estimators that wrap other estimators:
predict_proba can be a property in wrapped estimator but a method
in wrapper estimator, so the AttributeError will be raised in
runtime. But we don't want to catch it in runtime because it might
be unrelated. This fixes a failure that appeared after
https://github.com/scikit-learn-contrib/lightning/pull/98/ merge
(see https://travis-ci.org/TeamHG-Memex/eli5/jobs/172586529)