explain_prediction with pandas: xgboost and sklearn
Convert dataframe row (Series) to a dataframe
with a single element.
XGBoost expects feature names to stay the
same, so you can't train it on a dataframe and then pass a numpy
array: there will be a feature name mismatch. This is why we need
to make it a dataframe, not convert to a 2d numpy array.